fix: no plots is rendered
Due to the introduction of width and height parameters in the following
py-shiny commit,
https://github.com/posit-dev/py-shiny/commit/fee9752d5d9e507a7f9fc6751deec1290190290f
the plots in scrollable body all have 0 height and lead to linear
algebra errors when `matplotlib` attempt to apply tight layout.

The solution is to manually specify the height to `render.plot` and to
prevent the parent `div` of `img` having 0 height, add a CSS style that
specifies auto minimum height for `div`'s with `img` in it.